+++ OnlySnarf/__init__.py
@@ -1,0 +1,2 @@
+from .snarf import Snarf
+from .menu import Menu