+++ bin/start-api.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+export FLASK_APP=./api/index.py
+pipenv run flask --debug run -h 0.0.0.0